[Java] Test cluster session order is maintained through recovery.
`ClusteredService` implementations may accidentally rely on the order of
sessions in `Cluster#clientSessions`. Currently, the iteration order
may not be the same across nodes, e.g., if they have recovered from
different snapshots due to a restart, and this difference could lead to
divergence.

`Cluster#clientSessions` is backed by a `Long2ObjectHashMap`. The
differences in order arise due to different map capacities on
nodes that have experienced different snapshots and sequences of session
open/close events.

In this commit, I have added two tests. Each test computes a checksum by
iterating over the sessions before and after recovery and expects these
to be equal. When recovering without a snapshot, the map capacity is the
same during recovery and the test passes. When recovering from a
snapshot, the map capacity is different and the test fails.