Show a series of unfortunate events leading to duplication
Adds a test that shows a ~ 2 second pause on a client conductor can
lead to message duplication.

Scenario:
- one driver
- two Aeron clients:
  - a publisher uses client A
  - a subscriber uses client B
- the publisher establishes an IPC publication
- the subscriber establishes a subscription and waits for an image
- the publisher publishes a single message
- the subscriber polls the image and processes a single message
- [STALL START] the subscriber's client conductor stalls for ~ 2 seconds
- the publisher closes its publication
- the publisher does something to reuse counters (e.g., creating a new
publication)
- the subscriber polls and sees the message again
- [STALL ENDS]

Background:

I recently worked on some improvements to some tooling where we wanted
to give an IPC replay publication's log buffer time to be freed before
attempting another replay. The aim was to avoid overwhelming `/dev/shm`.
In this work, I find the `pub-pos` counter associated with the replay
image's `sub-pos` counter, then I periodically check that the `pub-pos`
counter is still there. It wasn't a perfect fix, as the `pub-pos`
counter is closed just _before_ the logbuffer is freed.

As part of this change, I had to consider the conditions under which I
could safely navigate this relationship between `sub-pos` counter and
`pub-pos` counter, which got me thinking about lifetimes of these
counters.